fix: clean up manta-pay arkworks directory organization
